Ported everything to typescript and removed some stuff that wasn't being used like the Ui code
