+++ gulpfile.js
@@ -1,7 +1,6 @@
 const { series, src, dest, pipe } = require('gulp');
 const { exec } = require('child_process')
 const fs = require('fs');
-const path = require('path');
 const rimraf = require('rimraf');
 const ts = require('gulp-typescript');
 const ncp = require('ncp').ncp;
@@ -13,11 +12,16 @@ const tsProject = ts.createProject('tsconfig.json')
 
 function clean(cb){
 
-    fs.existsSync(outputPath) && (
+    const deleteDist = () => {
         rimraf(outputPath, (err)=>{
-            err && cb(new Error('Could not remove ./dist folder'));
+            if (err){
+                console.log('Error deleting ./dist folder :', err);
+            }
         })
-    )
+        console.log('Deleted dist folder')
+    }
+
+    if (fs.existsSync(outputPath)) deleteDist()
 
     cb();
 };
@@ -34,15 +38,13 @@ function transpile(){
 function copyFiles(cb){
     // Use double quotes for include (and this whole function to be safe) for Windows compatibility
 
-    const excludeList = [
-        "*.md",
-        "*.ts"
-    ];
-    
-    //TODO: Write include function https://stackoverflow.com/questions/45842768/node-ncp-filter-not-working
     
     const filter = (file) => {
-
+        
+        const excludeList = [
+            "*.md",
+            "*.ts"
+        ];
         let relativeSliceIndex = file.indexOf("/src");
         let relativePath = "." + file.substring(relativeSliceIndex, file.length);
         
@@ -59,23 +61,26 @@ function copyFiles(cb){
 
     }
     ncp(inputPath, outputPath, options, (err) => {
-        err ? cb(new Error('Error copying files')) : cb();
+        if (err) {
+          cb(new Error());
+        }
+        cb();
     });
-    
+
 };
 
 function generateFiles(cb){
     //TODO: Change this over to the programmatic API
     exec(' npx typescript-json-schema ./src/config/configInterface.ts Config --out ./src/config/configSchema.json', (err, stdout, stderr)=>{
         if (err){ 
-            console.log(err);
+            console.log("HELLO")
             cb(new Error('Error generating JSON Schema from Typescript interface.'));
         };
     });
     cb();
 }
 
-const build = series(clean, generateFiles, transpile, copyFiles)
+const build = series(clean, generateFiles, copyFiles, transpile);
 
 module.exports = {
     clean,
