--- src/lib/engine.js
@@ -1,5 +1,0 @@
-
-
-module.exports = class Engine {
-    
-}
