--- src/utils/index.js
@@ -1,5 +1,0 @@
-const kill = require('./kill');
-
-module.exports = {
-    kill,
-}
