+++ src/utils/index.ts
@@ -1,0 +1,5 @@
+import kill from './kill';
+
+export {
+    kill
+}
