--- src/utils/kill.js
@@ -1,4 +1,0 @@
-module.exports = function kill(message = "Exiting Algotia...", signal = "SIGINT") {
-    console.log(message);
-    process.kill(process.pid, signal);
-};
