+++ src/utils/kill.ts
@@ -1,0 +1,4 @@
+export default function kill(message: string = "Exiting Algotia...", signal: string | number = "SIGINT") {
+    console.log(message);
+    process.kill(process.pid, signal);
+};
