+++ tsconfig.json
@@ -1,6 +1,6 @@
 {
   "compilerOptions": {
-    "target": "esnext",                          
+    "target": "ES2016",                          
     "module": "commonjs",                     
     "allowJs": true,                       
     "outDir": "./dist",
